fix ubi create size overflow

+++ common/cmd_ubi.c
@@ -164,7 +164,7 @@ bad:
 	return err;
 }
 
-static int ubi_create_vol(char *volume, int size, int dynamic)
+static int ubi_create_vol(char *volume, unsigned long size, int dynamic)
 {
 	struct ubi_mkvol_req req;
 	int err;
@@ -188,7 +188,7 @@ static int ubi_create_vol(char *volume, int size, int dynamic)
 		printf("verify_mkvol_req failed %d\n", err);
 		return err;
 	}
-	printf("Creating %s volume %s of size %d\n",
+	printf("Creating %s volume %s of size %u\n",
 		dynamic ? "dynamic" : "static", volume, size);
 	/* Call real ubi create volume */
 	return ubi_create_volume(ubi, &req);
@@ -433,7 +433,7 @@ static int ubi_dev_scan(struct mtd_info *info, char *ubidev,
 
 static int do_ubi(cmd_tbl_t * cmdtp, int flag, int argc, char * const argv[])
 {
-	size_t size = 0;
+	unsigned long size = 0;
 	ulong addr = 0;
 	int err = 0;
